+++ test/unit/types.ts
@@ -8,6 +8,7 @@ export type RxTestStorage = {
     // TODO remove name here, it can be read out already via getStorage().name
     readonly name: string;
     readonly getStorage: () => RxStorage<any, any>;
+    readonly hasMultiInstance: boolean;
     readonly hasCouchDBReplication: boolean;
     readonly hasAttachments: boolean;
     // true if the storage supports $regex queries, false if not.